+++ qase-javascript-commons/src/index.ts
@@ -8,3 +8,4 @@ export * from './reporters';
 export * from './writer';
 
 export * from './utils/get-package-version';
+export * from './utils/mimeTypes';