+++ qase-javascript-commons/src/models/test-result.ts
@@ -2,21 +2,6 @@ import { TestStepType } from './test-step';
 import { Attachment } from './attachment';
 import { TestExecution } from './test-execution';
 
-
-// export type TestResultType = {
-//   id: string;
-//   testOpsId: number[];
-//   title: string;
-//   status: `${TestStatusEnum}`;
-//   suiteTitle?: string | string[] | undefined;
-//   error?: Error | undefined;
-//   startTime?: number | undefined;
-//   duration?: number | undefined;
-//   endTime?: number | undefined;
-//   steps?: TestStepType[] | undefined;
-//   attachments?: string[] | undefined;
-// };
-
 export type TestResultType = {
   id: string
   title: string